+++ modules/moderation/report.py
@@ -131,14 +131,15 @@ class Report(cogs.BaseCog):
             await interaction.response.send_message(embed=user_embed, ephemeral=True)
             return
 
-        role = interaction.guild.get_role(
-            int(
-                configuration.get_config_entry(interaction.guild.id, "report_ping_role")
-            )
+        report_ping_role = configuration.get_config_entry(
+            interaction.guild.id, "report_ping_role"
         )
 
+        if report_ping_role:
+            report_ping_role = f"<@&{report_ping_role}>"
+
         await alert_channel.send(
-            content=role.mention,
+            content=report_ping_role,
             embed=embed,
             allowed_mentions=discord.AllowedMentions(roles=True),
         )